refined the session manager

+++ public/js/services/SessionManager.js
@@ -10,14 +10,20 @@
     *   @dependency {Angular} $q
     *   @dependency {Present} logger -- configurable logger for development
     *   @dependency {Present} UserContextApiClient -- interacts directly with the User Contexts Api Client
-    *   @dependency {Present} ApiClientResponseHandler -- handles the raw api response
     */
 
-    return PServices.factory('SessionManager', ['$q', 'localStorageService', 'logger', 'UserContextApiClient', 'ApiClientResponseHandler',
+    return PServices.factory('SessionManager', ['$q', 'localStorageService', 'logger', 'UserContextApiClient',
 
       function($q, localStorageService, logger, UserContextApiClient) {
 
         return {
+
+          /**
+           * createNewSession
+           * Sends a request to create a new user context and stores it to local storage on success
+           *   @param <String> username -- username in which the user context will be created with
+           *   @param <String> password -- password to validate the user
+           */
 
           createNewSession : function(username, password) {
 
@@ -25,18 +31,24 @@
 
             UserContextApiClient.createNewUserContext(username, password)
               .then(function(rawApiResponse) {
-                logger.debug(['PServices.SessionManager.login -- creating new session token'], rawApiResponse);
+                logger.debug(['PServices.SessionManager.creatingNew Session -- creating new session token'], rawApiResponse);
                 localStorageService.set('sessionToken', rawApiResponse.result.object.sessionToken);
                 localStorageService.set('userId', rawApiResponse.result.object.user.object._id);
                 creatingSession.resolve();
               })
               .catch(function() {
+                logger.error(['PServices.SessionManager.creatingNewSession -- couldn\'t create session token'])
                 creatingSession.reject();
               });
 
             return creatingSession.promise
 
           },
+
+          /**
+           * destroyCurrentSession
+           * Sends a request to delete the user context and clears session token from local storage
+           */
 
           destroyCurrentSession : function() {
 
@@ -47,22 +59,32 @@
               userId: localStorageService.get('userId')
             };
 
-            logger.debug(['PServices.SessionManager.destroyCurrentSession -- destroying user session']);
+            if(session.token && session.userId) {
 
-            UserContextApiClient.destroyUserContext(session)
-               .then(function() {
-                 localStorageService.clearAll();
-                 deletingSession.resolve();
-               })
-               .catch(function() {
-                 localStorageService.clearAll();
-                 logger.error(['PServices.SessionManager.destroyCurrentSession -- user context deletion failed, session data being deleted regardless']);
-                 deletingSession.resolve();
-               });
+              UserContextApiClient.destroyUserContext(session)
+                .then(function() {
+                  logger.debug(['PServices.SessionManager.destroyCurrentSession', 'User context deleted. Destroying current session']);
+                  localStorageService.clearAll();
+                  deletingSession.resolve();
+                })
+                .catch(function() {
+                  logger.error(['PServices.SessionManager.destroyCurrentSession', 'User context deletion failed, session data being deleted regardless']);
+                  localStorageService.clearAll();
+                  deletingSession.reject();
+                });
+            } else {
+              logger.error(['PServices.SessionManager.destroyCurrentSession -- no session set. ']);
+              deletingSession.reject();
+            }
 
             return deletingSession.promise;
 
           },
+
+          /**
+           * getCurrentSession
+           * Returns the session token if it exists. Returns false if the session token is invalid
+           */
 
           getCurrentSession : function() {
 
@@ -71,21 +93,8 @@
               userId: localStorageService.get('userId')
             };
 
-            return session;
-
-          },
-
-          checkForValidSession : function() {
-
-            var session = {
-              token: localStorageService.get('sessionToken'),
-              userId: localStorageService.get('userId')
-            };
-
-            if(session.token && session.userId) return true
+            if(session.token && session.userId) return session;
             else return false;
-
-            logger.test(['PServices.SessionManager -- checking for valid session', session]);
 
           }
